--- pkg/planner/core/casetest/physicalplantest/BUILD.bazel
@@ -1,36 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_test")
-
-go_test(
-    name = "physicalplantest_test",
-    timeout = "short",
-    srcs = [
-        "main_test.go",
-        "physical_plan_test.go",
-    ],
-    data = glob(["testdata/**"]),
-    flaky = True,
-    race = "on",
-    shard_count = 33,
-    deps = [
-        "//pkg/config",
-        "//pkg/domain",
-        "//pkg/executor",
-        "//pkg/infoschema",
-        "//pkg/parser",
-        "//pkg/parser/ast",
-        "//pkg/parser/model",
-        "//pkg/planner",
-        "//pkg/planner/core",
-        "//pkg/planner/util/coretestsdk",
-        "//pkg/testkit",
-        "//pkg/testkit/external",
-        "//pkg/testkit/testdata",
-        "//pkg/testkit/testmain",
-        "//pkg/testkit/testsetup",
-        "//pkg/util/context",
-        "//pkg/util/hint",
-        "@com_github_pingcap_failpoint//:failpoint",
-        "@com_github_stretchr_testify//require",
-        "@org_uber_go_goleak//:goleak",
-    ],
-)
